update option picker
