+++ lib/screens/onboarding/signup/onboarding_basic_info_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:solarisdemo/config.dart';
+import 'package:solarisdemo/models/select_option.dart';
 import 'package:solarisdemo/redux/app_state.dart';
 import 'package:solarisdemo/redux/onboarding/signup/onboarding_signup_action.dart';
 import 'package:solarisdemo/screens/onboarding/signup/onboarding_email_screen.dart';
