+++ lib/widgets/ivory_select_option.dart
@@ -1,9 +1,9 @@
-import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:solarisdemo/config.dart';
-import 'package:solarisdemo/utilities/debouncer.dart';
-import 'package:solarisdemo/widgets/custom_builder.dart';
-import 'package:solarisdemo/widgets/ivory_text_field.dart';
+import 'package:solarisdemo/models/select_option.dart';
+
+import 'package:solarisdemo/widgets/ivory_option_picker.dart';
+
 import 'package:solarisdemo/widgets/modal.dart';
 
 class IvorySelectOption extends StatefulWidget {
@@ -19,7 +19,6 @@ class IvorySelectOption extends StatefulWidget {
   final VoidCallback? onBottomSheetOpened;
   final bool filterOptions;
   final bool bottomSheetExpanded;
-  final Widget Function(BuildContext, SelectOption)? optionSeparatorBuilder;
   final String searchFieldInitialText;
   final bool statusbarVisibilityForTallModal;
 
@@ -38,7 +37,6 @@ class IvorySelectOption extends StatefulWidget {
     this.onSearchChanged,
     this.filterOptions = true,
     this.bottomSheetExpanded = false,
-    this.optionSeparatorBuilder,
     this.statusbarVisibilityForTallModal = false,
   });
 
@@ -105,7 +103,7 @@ class _IvorySelectOptionState extends State<IvorySelectOption> {
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
                     children: [
                       if (prefixItems.isNotEmpty) ...[
-                        ...prefixItems,
+                        ...prefixItems.take(3).toList(),
                         const SizedBox(width: 8),
                       ],
                       if (_controller.selectedOptions.isNotEmpty)
@@ -167,9 +165,7 @@ class _IvorySelectOptionState extends State<IvorySelectOption> {
         onSearchChanged: widget.onSearchChanged,
         filterOptions: widget.filterOptions,
         expanded: widget.bottomSheetExpanded,
-        optionSeparatorBuilder: widget.optionSeparatorBuilder,
         onOptionSelected: (option) {
-          _controller.selectOption(option);
           widget.onOptionSelected?.call(option);
         },
       ),
@@ -177,217 +173,15 @@ class _IvorySelectOptionState extends State<IvorySelectOption> {
   }
 }
 
-class IvoryOptionPicker extends StatefulWidget {
-  final IvorySelectOptionController controller;
-  final String searchFieldPlaceholder;
-  final void Function(SelectOption) onOptionSelected;
-  final void Function(String)? onSearchChanged;
-  final Widget Function(BuildContext, SelectOption)? optionSeparatorBuilder;
-  final bool enabledSearch;
-  final bool filterOptions;
-  final bool expanded;
-  final String searchFieldInitialText;
-
-  const IvoryOptionPicker({
-    super.key,
-    required this.controller,
-    required this.onOptionSelected,
-    required this.enabledSearch,
-    required this.searchFieldPlaceholder,
-    required this.filterOptions,
-    this.onSearchChanged,
-    this.expanded = false,
-    this.optionSeparatorBuilder,
-    this.searchFieldInitialText = "",
-  });
-
-  @override
-  State<IvoryOptionPicker> createState() => _IvoryOptionPickerState();
-}
-
-class _IvoryOptionPickerState extends State<IvoryOptionPicker> {
-  late List<SelectOption> _filteredOptions;
-  final _debouncer = Debouncer(milliseconds: 500);
-
-  @override
-  void initState() {
-    super.initState();
-
-    _filteredOptions = widget.controller.options;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return CustomBuilder(
-      builder: (BuildContext context, child) {
-        if (widget.expanded) {
-          return Expanded(child: child!);
-        }
-
-        return child!;
-      },
-      child: Column(
-        children: [
-          if (widget.enabledSearch) ...[
-            Padding(
-              padding: ClientConfig.getCustomClientUiSettings().defaultScreenHorizontalPadding,
-              child: IvoryTextField(
-                placeholder: widget.searchFieldPlaceholder,
-                initialText: widget.searchFieldInitialText,
-                suffix: Icon(Icons.search, color: ClientConfig.getCustomColors().neutral700, size: 20),
-                onChanged: (value) {
-                  widget.onSearchChanged?.call(value);
-
-                  if (widget.filterOptions) {
-                    _debouncer.run(
-                      () => setState(() {
-                        _filteredOptions = widget.controller.options
-                            .where((option) => option.textLabel.toLowerCase().contains(value.toLowerCase()))
-                            .toList();
-                      }),
-                    );
-                  }
-                },
-              ),
-            ),
-            const SizedBox(height: 24),
-          ],
-          widget.filterOptions
-              ? _buildListView(options: _filteredOptions)
-              : ListenableBuilder(
-                  listenable: widget.controller,
-                  builder: (context, child) {
-                    if (widget.controller.loading) {
-                      return Center(
-                        child: CircularProgressIndicator(
-                          strokeWidth: 3,
-                          color: ClientConfig.getCustomColors().neutral700,
-                        ),
-                      );
-                    }
-
-                    return _buildListView(options: _filteredOptions);
-                  },
-                ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildListView({required List<SelectOption> options}) {
-    if (options.isEmpty) {
-      return Align(
-        alignment: Alignment.topCenter,
-        child: Text(
-          "No results found",
-          style: ClientConfig.getTextStyleScheme().bodyLargeRegularBold,
-        ),
-      );
-    }
-
-    return CustomBuilder(
-      builder: (context, child) {
-        if (widget.expanded) {
-          return Expanded(child: child!);
-        }
-
-        return child!;
-      },
-      child: ListView.separated(
-        shrinkWrap: true,
-        itemCount: options.length,
-        separatorBuilder: (context, index) {
-          final SelectOption option = options[index];
-
-          if (widget.optionSeparatorBuilder != null) {
-            return widget.optionSeparatorBuilder!(context, option);
-          }
-
-          return const SizedBox();
-        },
-        itemBuilder: (context, index) {
-          SelectOption option = options[index];
-
-          return _BottomSheetOption(
-            key: UniqueKey(),
-            textLabel: option.textLabel,
-            isSelected: option.selected,
-            multiselect: widget.controller.multiselect,
-            onTap: () {
-              widget.onOptionSelected(option);
-            },
-            prefix: option.prefix,
-          );
-        },
-      ),
-    );
-  }
-}
-
-class _BottomSheetOption extends StatefulWidget {
-  final String textLabel;
-  final bool isSelected;
-  final bool multiselect;
-  final VoidCallback onTap;
-  final Widget? prefix;
-
-  const _BottomSheetOption({
-    super.key,
-    required this.textLabel,
-    required this.isSelected,
-    required this.onTap,
-    this.multiselect = false,
-    this.prefix,
-  });
-
-  @override
-  State<_BottomSheetOption> createState() => _BottomSheetOptionState();
-}
-
-class _BottomSheetOptionState extends State<_BottomSheetOption> {
-  late bool _isSelected;
-
-  @override
-  void initState() {
-    super.initState();
-    _isSelected = widget.isSelected;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        widget.onTap();
-        if (widget.multiselect) {
-          setState(() {
-            _isSelected = !_isSelected;
-          });
-        } else {
-          Navigator.pop(context);
-        }
-      },
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 16),
-        color: _isSelected ? ClientConfig.getCustomColors().neutral100 : null,
-        child: Row(
-          children: [
-            if (widget.prefix != null) widget.prefix!,
-            Text(widget.textLabel, style: ClientConfig.getTextStyleScheme().heading4),
-            const Spacer(),
-            if (_isSelected) Icon(Icons.check, color: ClientConfig.getCustomColors().success),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
 class IvorySelectOptionController extends ChangeNotifier {
-  final bool multiselect;
   bool _loading;
   bool _enabled;
+  bool multiselect;
+  String _searchText = "";
 
   final List<SelectOption> _options;
+  final List<SelectOption> _initialOptions;
+  final List<SelectOption> _selectedOptions = List.empty(growable: true);
 
   IvorySelectOptionController({
     this.multiselect = false,
@@ -395,26 +189,79 @@ class IvorySelectOptionController extends ChangeNotifier {
     bool enabled = true,
     List<SelectOption>? options,
   })  : _options = options ?? List.empty(growable: true),
+        _initialOptions = options ?? List.empty(growable: true),
         _loading = loading,
         _enabled = enabled;
 
   void setOptions(List<SelectOption> options) {
     _options.clear();
+    _initialOptions.clear();
+
     _options.addAll(options);
+    _initialOptions.addAll(options);
 
     notifyListeners();
   }
 
-  void selectOption(SelectOption selectedOption) {
-    for (int optionIndex = 0; optionIndex < _options.length; optionIndex++) {
-      SelectOption option = _options[optionIndex];
-      if (selectedOption.value == option.value) {
-        _options[optionIndex] = option.copyWith(selected: !option.selected);
-      } else {
-        _options[optionIndex] = multiselect ? option : option.copyWith(selected: false);
+  void toggleOptionSelection(SelectOption selectedOption, int index) {
+    if (selectedOption.selected) {
+      _selectedOptions.removeWhere((element) => element.value == selectedOption.value);
+
+      final optionIndex = _options.indexWhere((element) => element.value == selectedOption.value);
+      _options[optionIndex] = _options[optionIndex].copyWith(selected: false);
+    } else {
+      if (!multiselect) {
+        _selectedOptions.clear();
+        for (int optionIndex = 0; optionIndex < _options.length; optionIndex++) {
+          _options[optionIndex] = _options[optionIndex].copyWith(selected: false);
+        }
+      }
+
+      _selectedOptions.add(selectedOption.copyWith(selected: true));
+    }
+
+    _options.removeWhere((option) => selectedOptions.any((element) => element.value == option.value));
+    _options.insertAll(0, selectedOptions);
+
+    notifyListeners();
+  }
+
+  void filterOptionsByText(String searchText) {
+    _searchText = searchText;
+
+    final filtered = List<SelectOption>.empty(growable: true);
+    for (int optionIndex = 0; optionIndex < _initialOptions.length; optionIndex++) {
+      final option = _initialOptions[optionIndex];
+      final isSelected = selectedOptions.any((element) => element.value == option.value);
+
+      if (option.textLabel.toLowerCase().contains(searchText.toLowerCase())) {
+        filtered.add(option.copyWith(selected: isSelected));
       }
     }
 
+    _options.clear();
+    _options.addAll(filtered);
+
+    notifyListeners();
+  }
+
+  void resetFilter() {
+    _searchText = "";
+
+    final options = List<SelectOption>.empty(growable: true);
+    for (int optionIndex = 0; optionIndex < _initialOptions.length; optionIndex++) {
+      final option = _initialOptions[optionIndex];
+      final isSelected = selectedOptions.any((element) => element.value == option.value);
+
+      if (isSelected) {
+        options.insert(0, option.copyWith(selected: true));
+      } else {
+        options.add(option);
+      }
+    }
+
+    _options.clear();
+    _options.addAll(options);
     notifyListeners();
   }
 
@@ -438,37 +285,7 @@ class IvorySelectOptionController extends ChangeNotifier {
 
   bool get loading => _loading;
   bool get enabled => _enabled;
+  String get searchText => _searchText;
   List<SelectOption> get options => _options;
-  List<SelectOption> get selectedOptions => _options.where((option) => option.selected).toList();
-}
-
-class SelectOption extends Equatable {
-  final Widget? prefix;
-  final String textLabel;
-  final bool selected;
-  final String value;
-
-  const SelectOption({
-    required this.value,
-    required this.textLabel,
-    this.selected = false,
-    this.prefix,
-  });
-
-  SelectOption copyWith({
-    String? value,
-    String? textLabel,
-    bool? selected,
-    Widget? prefix,
-  }) {
-    return SelectOption(
-      value: value ?? this.value,
-      textLabel: textLabel ?? this.textLabel,
-      selected: selected ?? this.selected,
-      prefix: prefix ?? this.prefix,
-    );
-  }
-
-  @override
-  List<Object?> get props => [value, textLabel, selected, prefix];
+  List<SelectOption> get selectedOptions => _selectedOptions;
 }
